config.py
import os
import re
from dotenv import load_dotenv
import logging
from typing import Dict, List, Pattern

# Настройка логгирования
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    handlers=[
        logging.FileHandler("video_bot.log"),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

load_dotenv('token.env')
load_dotenv('.env', override=False)

# Основные настройки
BOT_TOKEN: str = os.getenv('BOT_TOKEN', '')
VK_ACCESS_TOKEN: str = os.getenv('VK_ACCESS_TOKEN', '')
VK_API_VERSION: str = '5.199'
DOWNLOAD_DIR: str = "downloads"
MAX_FILE_SIZE: int = int(os.getenv('MAX_FILE_SIZE', str(2 * 1024 * 1024 * 1024)))  # 2GB
SELENIUM_REMOTE_URL: str = os.getenv('SELENIUM_REMOTE_URL', '')
TWITTER_USERNAME: str = os.getenv('TWITTER_USERNAME', '')
TWITTER_PASSWORD: str = os.getenv('TWITTER_PASSWORD', '')
TELEGRAM_API_URL: str = os.getenv('TELEGRAM_API_URL', '').strip()

# Поддерживаемые платформы
PLATFORMS = {
    "yandex_zen": r"zen\.yandex\.ru|dzen\.ru",
    "youtube": r"(youtube\.com|youtu\.be)",
    "instagram": r"instagram\.com",
    "tiktok": r"tiktok\.com|vm\.tiktok\.com",
    "twitter": r"(x\.com|twitter\.com)",
    "vk": r"(vk\.com|vkvideo\.ru)",  # Объединенные паттерны
    "reddit": r"(reddit\.com|packaged-media\.redd\.it)",
    # "dzen": r"dzen\.ru/video/watch"
}

# Паттерны для определения типа контента
TWITTER_PATTERNS: List[str] = ['/status/', 'x.com/', 'twitter.com/']
VK_PATTERNS = [
    'vkvideo.ru/',
    'vkvideo.ru',
    'vk.com/video',
    'vk.com/clip',
    'vk.com/wall',
    'vkvideo.ru/video',
    '/video-',
    '/clip-',
    '/wall',
    'vkvideo.ru/video-',  # для ссылок вида video-XXXXX_YYYYY
    'vkvideo.ru/clip-'    # для ссылок вида clip-XXXXX_YYYYY
]
FFMPEG_PATH = "ffmpeg"
# Instagram Settings
INSTAGRAM_API_ENDPOINT = "https://apihut.in/api/download/videos"
USE_INSTAGRAM_API = True  # Set to True to use API instead of Instaloader
MAX_MERGED_VIDEO_SIZE = 50  # MB
INSTAGRAM_API_KEY = os.getenv('INSTAGRAM_API_KEY')  # If required by your AP
MAX_TELEGRAM_VIDEO_SIZE = int(os.getenv('MAX_TELEGRAM_VIDEO_SIZE', '2000'))  # MB (Telegram local limit)
MAX_RETRIES = 2  # Максимальное количество попыток
PHOTO_DURATION = 3  # Длительность фото в объединенном видео (сек)
# Настройки прокси
PROXY_SETTINGS = {
    'test_urls': [
        "https://www.instagram.com",
        "https://api.ipify.org?format=json",
        "https://www.google.com"
    ],
    'timeout': 15
}
os.makedirs(DOWNLOAD_DIR, exist_ok=True)
