requirements.txt
aiogram
yt-dlp
ffmpeg-python
python-dotenv
instaloader
selenium
webdriver_manager